renamed Feed to PostFeed
